requirements.txt
openai==0.27.6
pandas==1.4
numpy==1.24.3
python-dotenv==0.19.0
beautifulsoup4==4.9.3
gradio==3.30.0
import sqlite3
